Update all the samples.
(except clojure + scalatra, because those are broken, will fixed with later master merge.)

+++ samples/client/petstore/tizen/src/Pet.h
@@ -1,23 +1,7 @@
 /*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
  * Pet.h
  *
- * 
+ * A pet for sale in the pet store
  */
 
 #ifndef _Pet_H_
@@ -34,7 +18,7 @@ namespace Tizen {
 namespace ArtikCloud {
 
 
-/*! \brief 
+/*! \brief A pet for sale in the pet store
  *
  */
 
@@ -44,7 +28,7 @@ public:
 	 */
 	Pet();
 	Pet(char* str);
-	
+
 	/*! \brief Destructor.
 	 */
 	virtual ~Pet();
@@ -59,11 +43,11 @@ public:
 
 	/*! \brief Get 
 	 */
-	long getId();
+	long long getId();
 
 	/*! \brief Set 
 	 */
-	void setId(long  id);
+	void setId(long long  id);
 	/*! \brief Get 
 	 */
 	Category getCategory();
@@ -101,7 +85,7 @@ public:
 	void setStatus(std::string  status);
 
 private:
-	long id;
+	long long id;
 	Category category;
 	std::string name;
 	std::list <std::string>photoUrls;
@@ -109,7 +93,7 @@ private:
 	std::string status;
 	void __init();
 	void __cleanup();
-	
+
 };
 }
 }
